# src/main/webui/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    ConvertResultResponse:
      type: object
      required:
      - status
      - newAssetId
      - immichServerUrl
      - error
      properties:
        status:
          type: string
        newAssetId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
        immichServerUrl:
          type:
          - string
          - "null"
        error:
          type:
          - string
          - "null"
    Instant:
      type: object
      required:
      - epochSeconds
      - nanosecondsOfSecond
      properties:
        epochSeconds:
          type: integer
          format: int64
        nanosecondsOfSecond:
          type: integer
          format: int32
    JobStatus:
      type: object
      required:
      - threadId
      - state
      properties:
        threadId:
          type: integer
          format: int64
        state:
          $ref: "#/components/schemas/State"
        assetsFound:
          type: integer
          format: int64
        assetsQueued:
          type: integer
          format: int64
        error:
          type:
          - string
          - "null"
        startedAt:
          $ref: "#/components/schemas/Instant"
        endedAt:
          anyOf:
          - $ref: "#/components/schemas/Instant"
          - type: "null"
    QueuedAssetsResponse:
      type: object
      required:
      - assets
      - page
      - size
      - total
      - pageCount
      - hasPrev
      - hasNext
      - contentTypes
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/StagedAssetDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        pageCount:
          type: integer
          format: int32
        hasPrev:
          type: boolean
        hasNext:
          type: boolean
        contentTypes:
          type: array
          items:
            type: string
    StagedAssetDto:
      type: object
      required:
      - assetId
      - contentType
      - currentState
      properties:
        assetId:
          $ref: "#/components/schemas/UUID"
        contentType:
          type: string
        currentState:
          type: string
    State:
      type: string
      enum:
      - RUNNING
      - DONE
      - FAILED
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
    UserAddParams:
      type: object
      properties:
        id:
          type:
          - integer
          - "null"
          format: int32
        name:
          type: string
        immichServerUrl:
          type: string
        apiKey:
          type: string
      required:
      - name
      - immichServerUrl
      - apiKey
    UserDetails:
      type: object
      required:
      - userInfo
      - serverVersion
      - tagId
      properties:
        userInfo:
          $ref: "#/components/schemas/UserInfo"
        serverVersion:
          type:
          - string
          - "null"
        tagId:
          anyOf:
          - $ref: "#/components/schemas/UUID"
          - type: "null"
    UserInfo:
      type: object
      required:
      - apiKey
      - name
      - immichServerUrl
      properties:
        apiKey:
          type: string
        name:
          type: string
        immichServerUrl:
          type: string
        _id:
          type:
          - integer
          - "null"
          format: int32
        id:
          type: integer
          format: int32
paths:
  /api/users:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAddParams"
        required: true
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
      summary: Upsert User
      tags:
      - User Resource
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserInfo"
      summary: List Users
      tags:
      - User Resource
  /api/users/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetails"
      summary: User Details
      tags:
      - User Resource
  /api/users/{id}/assets/queued:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - required: false
        name: contentTypes
        in: query
        schema:
          type:
          - array
          - "null"
          items:
            type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
        required: true
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 50
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueuedAssetsResponse"
      summary: Queued Assets
      tags:
      - User Resource
  /api/users/{id}/assets/refresh:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
      summary: Refresh Assets
      tags:
      - User Resource
  /api/users/{id}/assets/status:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
      summary: Refresh Status
      tags:
      - User Resource
  /api/users/{id}/assets/{assetId}/convert:
    post:
      parameters:
      - name: assetId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertResultResponse"
      summary: Convert Asset
      tags:
      - User Resource
info:
  title: immichCompactor API
  version: 1.0-SNAPSHOT