# src/main/webui/src/generated/models/job-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * immichCompactor API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Instant } from './instant';
// May contain unused imports in some cases
// @ts-ignore
import type { State } from './state';

export interface JobStatus {
    'threadId': number;
    'state': State;
    'assetsFound'?: number;
    'assetsQueued'?: number;
    'error'?: string | null;
    'startedAt'?: Instant;
    'endedAt'?: Instant | null;
}