# src/main/webui/src/generated/models/queued-assets-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * immichCompactor API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StagedAssetDto } from './staged-asset-dto';

export interface QueuedAssetsResponse {
    'assets': Array<StagedAssetDto>;
    'page': number;
    'size': number;
    'total': number;
    'pageCount': number;
    'hasPrev': boolean;
    'hasNext': boolean;
    'contentTypes': Array<string>;
}