# src/main/webui/src/generated/models/staged-asset-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * immichCompactor API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface StagedAssetDto {
    'assetId': string;
    'contentType': string;
    'currentState': string;
}